% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{authGoogle}
\alias{authGoogle}
\title{Google Drive authentication}
\usage{
authGoogle(tryToken = NULL, tryEmail = NULL)
}
\arguments{
\item{tryToken}{the shorthand name for the token, corresponding to the token filename
(e.g., if filename is \code{exampleproject-123456-abcdef789012.json},
use \code{tryToken = "exampleproject"}).}

\item{tryEmail}{character string giving the user email address to use.}
}
\description{
Attempts authentication using Google Service Account Token if found,
falling back to the specified email address.
}
