% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.bcnrv.R
\name{bcnrvConfig}
\alias{bcnrvConfig}
\title{BC NRV project configuration class}
\description{
BC NRV project configuration class

BC NRV project configuration class
}
\details{
This extends the \code{projConfig} class by setting various BC NRV config defaults,
and implements custom validation and finalizer methods.
}
\note{
See note in \code{?projConfig} describing the list-update mechanism of assignment to
certain fields.
}
\section{Super class}{
\code{\link[SpaDES.config:projConfig]{SpaDES.config::projConfig}} -> \code{bcnrvConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bcnrvConfig-new}{\code{bcnrvConfig$new()}}
\item \href{#method-bcnrvConfig-update}{\code{bcnrvConfig$update()}}
\item \href{#method-bcnrvConfig-clone}{\code{bcnrvConfig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SpaDES.config" data-topic="projConfig" data-id="validate"><a href='../../SpaDES.config/html/projConfig.html#method-projConfig-validate'><code>SpaDES.config::projConfig$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvConfig-new}{}}}
\subsection{Method \code{new()}}{
Create an new \code{bcnrvConfig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvConfig$new(projectPath, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectPath}}{character string giving the path to the project directory.}

\item{\code{...}}{Additional arguments passed to \code{useContext()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvConfig-update"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvConfig-update}{}}}
\subsection{Method \code{update()}}{
Update a \code{bcnrvConfig} object from its context.
Must be called anytime the context is updated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvConfig$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
