% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.bcnrv.R
\name{bcnrvContext}
\alias{bcnrvContext}
\title{BC NRV project context class}
\description{
BC NRV project context class

BC NRV project context class
}
\details{
This extends the \code{projContext} class by setting various defaults for BC NRV
and employing custom field validation.
}
\section{Super class}{
\code{\link[SpaDES.config:projContext]{SpaDES.config::projContext}} -> \code{bcnrvContext}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mode}}{Character string giving the project run mode.
One of 'production', 'development', or 'postprocess'.}

\item{\code{studyAreaHash}}{Character string giving the hash of current study area (read-only).}

\item{\code{studyAreaName}}{Character string giving the name of current study area.}

\item{\code{rep}}{replicate id (integer)}

\item{\code{pixelSize}}{raster pixel resolution (in metres) to use for simulations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bcnrvContext-new}{\code{bcnrvContext$new()}}
\item \href{#method-bcnrvContext-print}{\code{bcnrvContext$print()}}
\item \href{#method-bcnrvContext-clone}{\code{bcnrvContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvContext$new(
  projectPath,
  mode = "development",
  rep = 1L,
  res = 125,
  studyAreaName = "Chine"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectPath}}{Character string giving the path to the project directory.}

\item{\code{mode}}{Character string. One of 'production', 'development', or 'postprocess'.}

\item{\code{rep}}{Integer denoting the replicate ID for the current run.}

\item{\code{res}}{Numeric indicating the map resolution (pixel size) to use.
Must be one of 50, 125 (default), 250.}

\item{\code{studyAreaName}}{Character string identifying a study area (see \code{BC_HRV_preamble}
module for up-to-date descriptions of each study area label).}

\item{\code{version}}{Integer. Shorthand denoting whether vegetation parameter forcings (\code{version = 2})
should be used as they were for the ca. 2018 runs.
Version 3 uses the default LandR Biomass parameters (i.e., no forcings).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvContext-print"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvContext-print}{}}}
\subsection{Method \code{print()}}{
print the context object in markdown table format,
and invisibly return this formatted table for use
e.g., when writing the context info to a file for humans.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvContext$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bcnrvContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bcnrvContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bcnrvContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
