% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dbConnCache}
\alias{dbConnCache}
\title{Establish database connection for Cache}
\usage{
dbConnCache(type = "sqlite")
}
\arguments{
\item{type}{character string giving the Cache type.
One of \code{"sqlite"} or \code{"postgresql"}.}
}
\description{
Establish database connection for Cache
}
