% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{delay_rnd}
\alias{delay_rnd}
\title{Delay subsequent code execution by random interval}
\usage{
delay_rnd(t = 1L:15L)
}
\arguments{
\item{t}{integer vector from which to randomly draw a time (in minutes) to delay for}
}
\description{
Delay subsequent code execution by random interval
}
