% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\docType{data}
\name{.baseCachePath}
\alias{.baseCachePath}
\alias{.baseDataCachePath}
\alias{.baseModulePath}
\alias{.baseInputPath}
\alias{.baseOutputPath}
\alias{.baseLogPath}
\title{Default project paths}
\format{
An object of class \code{character} of length 1.

An object of class \code{NULL} of length 0.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
.baseCachePath

.baseDataCachePath

.baseModulePath

.baseInputPath

.baseOutputPath

.baseLogPath
}
\description{
Default project paths
}
\keyword{internal}
