% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\name{.getRelativePath}
\alias{.getRelativePath}
\title{Similar to e.g., \code{fs::path_rel} but from the 'other end' of the path,
i.e., working from the right (end) of the path instead of the left (beginning).
This allows extracting the project-relative paths for a project in different
directories, which is useful when some project files (e.g., cache, inputs, outputs)
are stored on a network drive and symlinked from the project directory.}
\usage{
.getRelativePath(path, relativeToPath)
}
\description{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{## here, assume p0 is a symlink pointing to p1;
## starting from the end, each of these paths have `myProject/` in common,
## with `outputs/` being the result relative to this path.
p0 <- "~/GitHub/myProject/outputs" ## symlink to p1
p1 <- "/mnt/projects/myProject/outputs"
p2 <- "~/GitHub/myProject"
.getRelativePath(p0, p2) ## "outputs"
.getRelativePath(p1, p2) ## "outputs"
}\if{html}{\out{</div>}}
}
\keyword{internal}
