% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{.linkProject}
\alias{.linkProject}
\title{Symlink to project configs for tests}
\usage{
.linkProject(prjMod)
}
\arguments{
\item{prjMod}{character, specifying file path to project config module}
}
\value{
NULL (invisibly)
}
\description{
Symlink to project configs for tests
}
\keyword{internal}
