% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.needPkg}
\alias{.needPkg}
\title{Notify the user of a missing package}
\usage{
.needPkg(pkg, enforce = "stop")
}
\arguments{
\item{pkg}{character string of length 1 giving the name of the package}

\item{enforce}{character string; one of "message", "warning", "stop",
which determines how the user is notified: by a message,
a warning, or an error, respectively.}
}
\description{
Notify the user of a missing package
}
\examples{
\dontrun{
  .needPkg("raster", "message")
  .needPkg("raster", "warning")
  .needPkg("raster", "stop")
}
}
\keyword{internal}
