% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\name{findProjectPath}
\alias{findProjectPath}
\alias{findProjectName}
\title{Find the project root directory}
\usage{
findProjectPath(from_wd = TRUE)

findProjectName()
}
\arguments{
\item{from_wd}{logical indicating whether to fall back to current working directory.}
}
\value{
\code{findProjectPath} returns an absolute path;
\code{findProjectName} returns the basename of the path.
}
\description{
Searches from current working directory for an Rstudio project file or git repository,
falling back on using the current working directory if \code{from_wd = TRUE}.
}
