% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{getModule}
\alias{getModule}
\title{Mock version of \code{SpaDES.project::getModule()}}
\usage{
getModule(modules, modulePath, overwrite, verbose)
}
\arguments{
\item{modules}{character string giving the name, or potential GitHub source of a module
(e.g., \code{"Biomass_core"} or \code{"PredictiveEcology/Biomass_core@development"})}

\item{modulePath}{path to the submodule}

\item{overwrite}{not used}

\item{verbose}{not used}
}
\description{
Allows use of SpaDES modules that make nested module calls (i.e., submodules).
For projects using version control with \emph{git} submodules, this version
simply creates a symlink to the project's version of the module.
This ensures the entire project stays in sync with itself, and
doesn't require installation of an additional package.
}
