% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{initialize,projConfig-method}
\alias{initialize,projConfig-method}
\title{Initialize a \code{projConfig} object}
\usage{
\S4method{initialize}{projConfig}(
  .Object,
  args = list(),
  options = list(),
  params = list(),
  paths = list(),
  ...
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{args}{Named list of additional project arguments.}

\item{options}{Named list of R and R package options to be set.}

\item{params}{Named list of named lists specifying simulation parameters.
The names of the outermost list must correspond to modules
(and may also include \code{.global}).}

\item{paths}{Named list of paths, which should include (at minimum) the
the paths in \code{SpaDES.core::setPaths}.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
Given the name or the definition of a class, plus optionally data to be
included in the object, \code{new} returns an object from that class.
}
