% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.landrfs.R
\name{landrfsConfig}
\alias{landrfsConfig}
\title{LandR-fireSense project configuration class}
\description{
LandR-fireSense project configuration class

LandR-fireSense project configuration class
}
\details{
This extends the \code{projConfig} class by setting various LandR-fireSense config defaults,
and implements custom validation and finalizer methods.
}
\note{
See note in \code{?projConfig} describing the list-update mechanism of assignment to
certain fields.
}
\section{Super class}{
\code{\link[SpaDES.config:projConfig]{SpaDES.config::projConfig}} -> \code{landrfsConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-landrfsConfig-new}{\code{landrfsConfig$new()}}
\item \href{#method-landrfsConfig-update}{\code{landrfsConfig$update()}}
\item \href{#method-landrfsConfig-clone}{\code{landrfsConfig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SpaDES.config" data-topic="projConfig" data-id="validate"><a href='../../SpaDES.config/html/projConfig.html#method-projConfig-validate'><code>SpaDES.config::projConfig$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsConfig-new}{}}}
\subsection{Method \code{new()}}{
Create an new \code{landrfsConfig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsConfig$new(projectName, projectPath, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectName}}{character string of length 1 giving the name of the project.}

\item{\code{projectPath}}{character string giving the path to the project directory.}

\item{\code{...}}{Additional arguments passed to \code{useContext()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsConfig-update"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsConfig-update}{}}}
\subsection{Method \code{update()}}{
Update a \code{landrfsConfig} object from its context.
Must be called anytime the context is updated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsConfig$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
