% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.landrfs.R
\name{landrfsContext}
\alias{landrfsContext}
\title{LandR-fireSense project context class}
\description{
LandR-fireSense project context class

LandR-fireSense project context class
}
\details{
This extends the \code{projContext} class by setting various defaults for LandR-fireSense
and employing custom field validation.
}
\section{Super class}{
\code{\link[SpaDES.config:projContext]{SpaDES.config::projContext}} -> \code{landrfsContext}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mode}}{Character string giving the project run mode.
One of 'production', 'development', or 'postprocess'.}

\item{\code{climateGCM}}{Character strings giving the CMIP6 climate scenario GCM from ClimateNA.
E.g., \code{"CanESM5"} or \code{"CNRM-ESM2-1"}.}

\item{\code{climateSSP}}{Numeric CMIP climate scenario SSP. E.g., \code{370} or \code{585}.}

\item{\code{pixelSize}}{raster pixel resolution (in metres) to use for simulations}

\item{\code{rep}}{replicate id (integer)}

\item{\code{studyAreaName}}{Character string giving the name of current study area.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-landrfsContext-new}{\code{landrfsContext$new()}}
\item \href{#method-landrfsContext-print}{\code{landrfsContext$print()}}
\item \href{#method-landrfsContext-clone}{\code{landrfsContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsContext$new(
  projectPath,
  climateGCM = NA_character_,
  climateSSP = NA_character_,
  mode = "development",
  rep = 1L,
  res = 250,
  studyAreaName = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectPath}}{Character string giving the path to the project directory.}

\item{\code{climateGCM}}{Character strings giving the CMIP6 climate scenario GCM from ClimateNA.
E.g., \code{"CanESM5"} or \code{"CNRM-ESM2-1"}.}

\item{\code{climateSSP}}{Numeric CMIP climate scenario SSP. E.g., \code{370} or \code{585}.}

\item{\code{mode}}{Character string. One of 'production', 'development', or 'postprocess'.}

\item{\code{rep}}{Integer denoting the replicate ID for the current run.}

\item{\code{res}}{Numeric indicating the map resolution (pixel size) to use.
Must be one of 125 or 250 (default).}

\item{\code{studyAreaName}}{Character string identifying a study area.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsContext-print"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsContext-print}{}}}
\subsection{Method \code{print()}}{
print the context object in markdown table format,
and invisibly return this formatted table for use
e.g., when writing the context info to a file for humans.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsContext$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landrfsContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-landrfsContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landrfsContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
