% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.landweb.R
\name{landwebContext}
\alias{landwebContext}
\title{LandWeb project context class}
\description{
LandWeb project context class

LandWeb project context class
}
\details{
This extends the \code{projContext} class by setting various LandWeb defaults and
employing custom fied validation.
}
\section{Super class}{
\code{\link[SpaDES.config:projContext]{SpaDES.config::projContext}} -> \code{landwebContext}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mode}}{Character string giving the project run mode.
One of 'production', 'development', 'postprocess', or 'profile'.}

\item{\code{studyAreaName}}{Character string giving the name of current study area.}

\item{\code{rep}}{replicate id (integer)}

\item{\code{dispersalType}}{Character string describing the seed dispersal type to use.
One of 'default', 'aspen', 'high', 'none'.}

\item{\code{forceResprout}}{Logical}

\item{\code{friMultiple}}{Numeric indicating a factor by which to scale the fire return intervals}

\item{\code{pixelSize}}{raster pixel resolution (in metres) to use for simulations}

\item{\code{ROStype}}{Character string describing the scaling of the LandMine fire model's
'rate of spread' parameters.
One of 'default' (i.e., none), 'equal' (i.e., all 1), 'log'.}

\item{\code{succession}}{logical}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-landwebContext-new}{\code{landwebContext$new()}}
\item \href{#method-landwebContext-print}{\code{landwebContext$print()}}
\item \href{#method-landwebContext-clone}{\code{landwebContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landwebContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-landwebContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landwebContext$new(
  projectPath,
  mode = "development",
  rep = 1L,
  res = 250,
  studyAreaName = "random",
  version = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectPath}}{Character string giving the path to the project directory.}

\item{\code{mode}}{Character string. One of 'production', 'development', 'postprocess',
or 'profile'.}

\item{\code{rep}}{Integer denoting the replicate ID for the current run.}

\item{\code{res}}{Numeric indicating the map resolution (pixel size) to use.
Must be one of 50, 100, 250 (default).}

\item{\code{studyAreaName}}{Character string identifying a study area (see \code{LandWeb_preamble}
module for up-to-date descriptions of each study area label).}

\item{\code{version}}{Integer. Shorthand denoting whether vegetation parameter forcings (\code{version = 2})
should be used as they were for the ca. 2018 runs.
Version 3 uses the default LandR Biomass parameters (i.e., no forcings).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landwebContext-print"></a>}}
\if{latex}{\out{\hypertarget{method-landwebContext-print}{}}}
\subsection{Method \code{print()}}{
print the context object in markdown table format,
and invisibly return this formatted table for use
e.g., when writing the context info to a file for humans.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landwebContext$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-landwebContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-landwebContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{landwebContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
