% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modList}
\alias{modList}
\title{Recursively modify elements of a list}
\usage{
modList(..., keep.null = TRUE)
}
\arguments{
\item{...}{named lists (possibly empty) where the elements of the first list will
be updated based on the values of subsequent lists.}

\item{keep.null}{logical indicating whether \code{NULL} elements should be kept or deleted.}
}
\description{
Wrapper around \code{utils::modifyList()} that can deal with more than two lists.
Based on \code{Require::modifyList3}.
}
\seealso{
\code{\link[utils:modifyList]{utils::modifyList()}}
}
