% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\name{paths4spades}
\alias{paths4spades}
\title{Get paths needed for SpaDES simulations}
\usage{
paths4spades(paths)
}
\arguments{
\item{paths}{A named list of paths.}
}
\description{
Given a list of multiple project-related paths,
return the subset accepted by \code{SpaDES.core::setPaths()}.
}
