% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printRunInfo}
\alias{printRunInfo}
\title{Print current run information}
\usage{
printRunInfo(context)
}
\arguments{
\item{context}{A named list of run descriptors, including e.g.,
the current user and machine names, the study area name, and replicate id.}
}
\description{
Print (as a message) the current run context information.
}
