% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\name{projectPaths}
\alias{projectPaths}
\title{Default project paths}
\usage{
projectPaths(type = NULL)
}
\arguments{
\item{type}{character. one of 'cache', 'modules', 'inputs', 'outputs', 'log'.}
}
\value{
character representing the project's default \code{type} path of type.
if \code{type} is \code{NULL} (i.e., not specified), returns a named list of default paths.
}
\description{
Default project paths
}
