% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setProjectOptions}
\alias{setProjectOptions}
\title{Set options based on project config}
\usage{
setProjectOptions(config)
}
\arguments{
\item{config}{a \code{projConfig} object}
}
\value{
invoked for its side-effect of setting options
(invisibly returns a list of the previously set options)
}
\description{
Set options based on project config
}
