% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shim_get_module}
\alias{shim_get_module}
\title{Override \code{SpaDES.project::getModule()} to use the project's version of the module.}
\usage{
shim_get_module(modules, modulePath, overwrite, verbose)
}
\arguments{
\item{modules}{See \code{?SpaDES.project::getModule}.}

\item{modulePath}{See \code{?SpaDES.project::getModule}.}

\item{overwrite}{See \code{?SpaDES.project::getModule}.}

\item{verbose}{See \code{?SpaDES.project::getModule}.}
}
\value{
See \code{?SpaDES.project::getModule}.
}
\description{
Override \code{SpaDES.project::getModule()} to use the project's version of the module.
}
