% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.landweb.R
\name{updateLandWebConfig}
\alias{updateLandWebConfig}
\title{Update LandWeb project configuration based on context}
\usage{
updateLandWebConfig(config, context)
}
\arguments{
\item{config}{a \code{landwebConfig} class object}

\item{context}{a named list}
}
\description{
Update LandWeb project configuration based on context
}
