% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts-update.R
\name{updateLandWebContext}
\alias{updateLandWebContext}
\title{Update LandWeb project context}
\usage{
updateLandWebContext(context, config)
}
\arguments{
\item{context}{a LandWeb context object created by \code{useContext("LandWeb", ...)}}

\item{config}{a \code{landwebConfig} object}
}
\value{
an updated context object
}
\description{
Update LandWeb project context
}
