% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R
\name{updateOutputPath}
\alias{updateOutputPath}
\title{Update project config output path}
\usage{
updateOutputPath(config, runNameFun)
}
\arguments{
\item{config}{a \code{projConfig} object}

\item{runNameFun}{function that computes a run name, returning a character string.}
}
\description{
Update project config output path
}
