% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{useConfig}
\alias{useConfig}
\title{Use a pre-defined project run context}
\usage{
useConfig(projectName = NULL, projectPath = NULL, ...)
}
\arguments{
\item{projectName}{character string of length 1 giving the name of the project.
Can be one of "LandWeb",
else will return a default context.}

\item{projectPath}{character string giving the path to the project directory.}

\item{...}{additional context arguments used by individual projects.}
}
\value{
named list
}
\description{
Use a pre-defined project run context
}
\examples{
\dontrun{
## default generic project config
config <- useConfig("myProject", "path/to/myMproject")

## default LandWeb project config
config <- useConfig("LandWeb", "~/GitHub/LandWeb", mode = "development", version = 3)
}
}
