% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{useContext}
\alias{useContext}
\title{Use a pre-defined project run context}
\usage{
useContext(projectName = NULL, projectPath = NULL, ...)
}
\arguments{
\item{projectName}{character string of length 1 giving the name of the project.
Can be one of "LandWeb",
else will return a default context.}

\item{projectPath}{Character string giving the path to the project directory.}

\item{...}{additional context arguments used by individual projects.}
}
\value{
a \code{projContext} object.
}
\description{
Use a pre-defined project run context
}
\examples{
\dontrun{
## default generic context
useContext("myProject", "path/to/myProject")

## default LandWeb context
useContext("LandWeb", "~/GitHub/LandWeb",
           mode = "development", rep = 1L, studyAreaName = "LandWeb", version = 3)
}
}
