% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{user}
\alias{user}
\alias{machine}
\title{Identify user or machine}
\usage{
user(name = NULL)

machine(name = NULL)
}
\arguments{
\item{name}{Optional character string giving user or machine name to match.}
}
\value{
if \code{name} is non-\code{NULL}, returns a logical indicating whether
the current user/machine matches \code{name}.
Otherwise returns a character string with the value of the current user/machine.
}
\description{
Identify user or machine
}
