% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-paths.R, R/helpers.R
\name{.isAbsolutePath}
\alias{.isAbsolutePath}
\alias{.isRstudioServer}
\title{Imports from other packages}
\usage{
.isAbsolutePath(pathnames)

.isRstudioServer()
}
\description{
Unexported functions from other packages; used internally.
}
\note{
\code{quickPlot} and \code{reproducible} are both dependencies of \code{SpaDES.core}
and should therefore be available, but we've copied the functions here
to avoid explicitly adding dependencies.
}
\keyword{internal}
