% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{projConfig-class}
\alias{projConfig-class}
\title{Project config class}
\description{
Project config class
}
\section{Slots}{

\describe{
\item{\code{options}}{Named list of R and R package options to be set.}

\item{\code{params}}{Named list of named lists specifying simulation parameters.
The names of the outermost list must correspond to modules
(and may also include \code{.global}).}

\item{\code{paths}}{Named list of paths, which should include (at minimum) the
the paths in \code{SpaDES.core::setPaths}.}

\item{\code{args}}{Named list of additional project arguments.}

\item{\code{.Data}}{Not used.}
}}

\author{
Alex Chubaty and Eliot McIntire
}
