% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{projConfig}
\alias{projConfig}
\title{Project configuration class}
\description{
Project configuration class

Project configuration class
}
\author{
Alex Chubaty and Eliot McIntire
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{args}}{Named list of additional project arguments.}

\item{\code{modules}}{List of module names, which should correspond to the names in \code{params}.}

\item{\code{options}}{Named list of R and R package options to be set.}

\item{\code{params}}{Named list of named lists specifying simulation parameters.
The names of the outermost list must correspond to modules
(and may also include \code{.global}).}

\item{\code{paths}}{Named list of paths, which should include (at minimum) the
the paths in \code{SpaDES.core::setPaths}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-projConfig-new}{\code{projConfig$new()}}
\item \href{#method-projConfig-update}{\code{projConfig$update()}}
\item \href{#method-projConfig-validate}{\code{projConfig$validate()}}
\item \href{#method-projConfig-clone}{\code{projConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-projConfig-new}{}}}
\subsection{Method \code{new()}}{
Create an new \code{projConfig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{projConfig$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projConfig-update"></a>}}
\if{latex}{\out{\hypertarget{method-projConfig-update}{}}}
\subsection{Method \code{update()}}{
Update the fields in a \code{projConfig} object

The objects passed as arguments should match the fields of the \code{projConfig} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{projConfig$update(
  args = NULL,
  modules = NULL,
  options = NULL,
  params = NULL,
  paths = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{named list; see \link{projConfig}.}

\item{\code{modules}}{named list; see \link{projConfig}.}

\item{\code{options}}{named list; see \link{projConfig}.}

\item{\code{params}}{named list; see \link{projConfig}.}

\item{\code{paths}}{named list; see \link{projConfig}.}

\item{\code{methods}}{named list; see \link{projConfig}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projConfig-validate"></a>}}
\if{latex}{\out{\hypertarget{method-projConfig-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the values of fields in a \code{projConfig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{projConfig$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-projConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-projConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{projConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
