% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-config-package.R
\docType{package}
\name{SpaDES.config-package}
\alias{SpaDES.config}
\alias{SpaDES.config-package}
\title{SpaDES.config package}
\description{
Tools for defining and validating project configurations across multiple contexts (including user- and machine-specific settings). Includes default configuration settings for several 'SpaDES' project types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spades-config.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.config}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.config/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@nrcan-rncan.gc.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
}

}
