% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{useContext}
\alias{useContext}
\title{Use a pre-defined project run context}
\usage{
useContext(projectName = NULL, ...)
}
\arguments{
\item{projectName}{character string of length 1 giving the name of the project.
Can be one of "LandWeb",
else will return a minimal set.}

\item{...}{additional context arguments used by individual projects.}
}
\value{
named list
}
\description{
Use a pre-defined project run context
}
\examples{
## default generic context
useContext()

## default LandWeb context
useContext("LandWeb", mode = "development", version = 3)
}
